ub: cdma: fix wait cqe and unregister ras
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/ID41DN
CVE: NA

-----------------------------------------------------------

This patch fix wait cqe fail when interruption is aggregated,
cdma unregister ras and cdma log level change

+++ cdma_u_common.h
@@ -105,7 +105,7 @@ struct cdma_u_jetty_queue {
 	uint32_t max_sge_num;
 	bool flush_flag;
 	bool lock_free;
-	uint8_t cstm; /* sq ctrl flag */
+	uint32_t cstm; /* sq ctrl flag */
 };
 
 struct cdma_u_buf {
